Fix auth bug and rename tests
When deployed on a live server, the auth callback would always try to
hit localhost.  This update to passport.js removes the localhost
component which allows the callback to redirect back to whatever host
the site was accessed on.

We also rename the test directory to tests, and update package.json
accordingly.