Update schemas and fix validation
This commit updates all the schemas to ensure we have a unique index
where needed and that we also have set the necessary fields as
required.  We also add error handling to categories and companies
controllers so we can detect if we get an error attempting to save a
duplicate document.