Create initial model testing
This commit starts testing models, beginning with the category model.
We also update tests/test.js to drop the database before running tests
and specify _country when creating a Job which was previously causing
a test to fail.  Lastly, we update the category model itself to alter
the message returned when a required field is missing.