Add contact form
This commit adds a contact form, which sends a message upon submission.
The actual e-mail is sent using the nodemailer library, which requires
SMTP details to be available in secrets.yml.

We also move should to a dev dependency in package.json and lastly we
update app.js so that requests to / will load
jobsController.indexAction.

Closes issue #8