Add composite unique index
Currently, we have a unique index on name in category and company
models.  This works fine except when you have mutliple sites trying to
save the same companies or categories.  This commit updates both models
to use a composite unique index on name and _site.

Additionally, we update all the model tests to use unique names to
avoid errors resulting from these tests running in parallel.  This
does not resolve the issue 100% but does seem to reduce errors
significantly.

Closes #10