Do not reference Country model in Job
As it is unclear how to make a reference field in a model optional, we
have removed the _country reference in Job and now just save the
country's name instead.  While it is nice having the reference to
Country, the Country table really should be static and having the
country name embedded in the job means we have to do one less query
when the job is being shown.

Closes #14